create components folder and organize files
