+++ src/components/Greetings.js
